Create a FooAioStub in Python to aid aio typing
aio GRPC is currently really hard to type (see grpc-stubs and mypy_protobuf projects), due to the dynamic nature of the generated GRPC interface. To help out the type stubs, create a separate but identical Stub type that behaves the same as the (sync) Stub but can have different typing information.